requirements.txt
requests
astral
python-dotenv
google-generativeai
pytest
